refactoring: made components smaller and put some code in new components, changed api requests from context to own seperate service file
