+++ src/data/db.json
@@ -907,5 +907,143 @@
       "type": "psychic"
     }
   ],
-  "teams": []
+  "teams": [
+    {
+      "id": "1",
+      "name": "Team #2",
+      "team": [
+        {
+          "id": "5",
+          "name": "charizard",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/6.png",
+          "type": "fire / flying"
+        },
+        "",
+        "",
+        {
+          "id": "3",
+          "name": "charmander",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/4.png",
+          "type": "fire"
+        },
+        "",
+        ""
+      ]
+    },
+    {
+      "id": "2",
+      "name": "Team #3",
+      "team": [
+        {
+          "id": "4",
+          "name": "charmeleon",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/5.png",
+          "type": "fire"
+        },
+        {
+          "id": "5",
+          "name": "charizard",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/6.png",
+          "type": "fire / flying"
+        },
+        "",
+        "",
+        "",
+        ""
+      ]
+    },
+    {
+      "id": "3",
+      "name": "Team #4",
+      "team": [
+        {
+          "id": "8",
+          "name": "blastoise",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/9.png",
+          "type": "water"
+        },
+        "",
+        "",
+        "",
+        "",
+        ""
+      ]
+    },
+    {
+      "id": "4",
+      "name": "test test",
+      "team": [
+        {
+          "id": "2",
+          "name": "venusaur",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/3.png",
+          "type": "grass / poison"
+        },
+        {
+          "id": "5",
+          "name": "charizard",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/6.png",
+          "type": "fire / flying"
+        },
+        "",
+        "",
+        "",
+        ""
+      ]
+    },
+    {
+      "id": "5",
+      "name": "dssdfqsd",
+      "team": [
+        {
+          "id": "0",
+          "name": "bulbasaur",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/1.png",
+          "type": "grass / poison"
+        },
+        {
+          "id": "1",
+          "name": "ivysaur",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/2.png",
+          "type": "grass / poison"
+        },
+        "",
+        {
+          "id": "3",
+          "name": "charmander",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/4.png",
+          "type": "fire"
+        },
+        {
+          "id": "4",
+          "name": "charmeleon",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/5.png",
+          "type": "fire"
+        },
+        {
+          "id": "5",
+          "name": "charizard",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/6.png",
+          "type": "fire / flying"
+        }
+      ]
+    },
+    {
+      "id": "6",
+      "name": "Team #7",
+      "team": [
+        {
+          "id": "2",
+          "name": "venusaur",
+          "imageUrl": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/official-artwork/3.png",
+          "type": "grass / poison"
+        },
+        "",
+        "",
+        "",
+        "",
+        ""
+      ]
+    }
+  ]
 }
